pyproject.toml
[project]
name = "nocodb-to-gpt-via-api"
version = "0.0.4"
description = "Alaska Small Business Development Center"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "numpy>=2.2.3",
    "ollama>=0.4.7",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[tool.uv.workspace]
members = ["example-app"]
